// src/types/general.ts
export type Result<T> = {
	status: 'success',
	data: T
} | {
	status: 'error',
	error: {
		message: string,
		status: number
	}
}